backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.16.0
webdriver-manager>=4.0.0
beautifulsoup4==4.12.2
pandas>=2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
lxml>=5.0.0
